split admin tests
